pyproject.toml
[project]
name = "shipcheck-engine"
version = "0.1.0"
description = "ShipCheck 페르소나 시뮬레이션 실험 엔진"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "pandas>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
analysis = [
    "scipy>=1.12",
    "pingouin>=0.5",
    "matplotlib>=3.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
